refactor: move dagre layout engine into selkie-layout
