+++ crates/selkie-layout/src/dagre/internal/graph.rs
@@ -7,7 +7,7 @@
 
 use std::collections::{HashMap, HashSet};
 
-use crate::layout::NodeShape;
+use crate::NodeShape;
 
 /// A multigraph with compound node support for dagre layout
 #[derive(Debug, Clone)]
