+++ crates/selkie-layout/src/dagre/internal/parent_dummy_chains.rs
