+++ crates/selkie-layout/tests/public_api.rs
@@ -1,4 +1,6 @@
-use selkie_layout::{layout, LayoutDirection, LayoutEdge, LayoutGraph, LayoutNode, LayoutOptions};
+use selkie_layout::{
+    layout, LayoutDirection, LayoutEdge, LayoutError, LayoutGraph, LayoutNode, LayoutOptions,
+};
 
 #[test]
 fn lays_out_simple_graph_with_public_api() {
@@ -17,6 +19,8 @@ fn lays_out_simple_graph_with_public_api() {
     let a = result.get_node("A").expect("node A exists");
     let b = result.get_node("B").expect("node B exists");
 
+    assert_eq!(a.size(), (80.0, 40.0));
+
     let a_pos = a.position().expect("node A has a position");
     let b_pos = b.position().expect("node B has a position");
     assert!(b_pos.y > a_pos.y);
@@ -33,7 +37,51 @@ fn rejects_missing_edge_endpoint() {
     graph.add_edge(LayoutEdge::new("missing", "A", "B"));
 
     let error = layout(graph).expect_err("missing endpoint should fail");
-    assert!(error.to_string().contains("B"));
+    assert_eq!(
+        error,
+        LayoutError::MissingEdgeEndpoint {
+            edge: "missing".to_string(),
+            endpoint: "B".to_string()
+        }
+    );
+}
+
+#[test]
+fn rejects_edge_with_empty_sources() {
+    let mut graph = LayoutGraph::new("invalid");
+    graph.add_node(LayoutNode::new("A", 80.0, 40.0));
+
+    let mut edge = LayoutEdge::new("empty_source", "A", "A");
+    edge.sources.clear();
+    graph.add_edge(edge);
+
+    let error = layout(graph).expect_err("empty sources should fail");
+    assert_eq!(
+        error,
+        LayoutError::MissingEdgeEndpoint {
+            edge: "empty_source".to_string(),
+            endpoint: "<source>".to_string()
+        }
+    );
+}
+
+#[test]
+fn rejects_edge_with_empty_targets() {
+    let mut graph = LayoutGraph::new("invalid");
+    graph.add_node(LayoutNode::new("A", 80.0, 40.0));
+
+    let mut edge = LayoutEdge::new("empty_target", "A", "A");
+    edge.targets.clear();
+    graph.add_edge(edge);
+
+    let error = layout(graph).expect_err("empty targets should fail");
+    assert_eq!(
+        error,
+        LayoutError::MissingEdgeEndpoint {
+            edge: "empty_target".to_string(),
+            endpoint: "<target>".to_string()
+        }
+    );
 }
 
 #[test]
@@ -43,7 +91,7 @@ fn rejects_duplicate_node_ids() {
     graph.add_node(LayoutNode::new("A", 80.0, 40.0));
 
     let error = layout(graph).expect_err("duplicate node should fail");
-    assert!(error.to_string().contains("Duplicate node id"));
+    assert_eq!(error, LayoutError::DuplicateNodeId("A".to_string()));
 }
 
 #[test]
@@ -52,7 +100,7 @@ fn rejects_invalid_node_dimensions() {
     graph.add_node(LayoutNode::new("A", -1.0, 40.0));
 
     let error = layout(graph).expect_err("negative width should fail");
-    assert!(error.to_string().contains("Invalid layout value"));
+    assert!(matches!(error, LayoutError::InvalidValue(message) if message.contains("node width")));
 }
 
 #[test]
@@ -64,7 +112,9 @@ fn rejects_invalid_spacing() {
     graph.add_node(LayoutNode::new("A", 80.0, 40.0));
 
     let error = layout(graph).expect_err("NaN spacing should fail");
-    assert!(error.to_string().contains("Invalid layout value"));
+    assert!(
+        matches!(error, LayoutError::InvalidValue(message) if message.contains("node_spacing"))
+    );
 }
 
 #[test]
@@ -73,7 +123,9 @@ fn rejects_missing_parent() {
     graph.add_node(LayoutNode::new("A", 80.0, 40.0).with_parent("missing"));
 
     let error = layout(graph).expect_err("missing parent should fail");
-    assert!(error.to_string().contains("Invalid parent relationship"));
+    assert!(
+        matches!(error, LayoutError::InvalidParent(message) if message.contains("missing parent"))
+    );
 }
 
 #[test]
@@ -83,7 +135,9 @@ fn rejects_parent_cycle() {
     graph.add_node(LayoutNode::new("B", 80.0, 40.0).with_parent("A"));
 
     let error = layout(graph).expect_err("parent cycle should fail");
-    assert!(error.to_string().contains("Invalid parent relationship"));
+    assert!(
+        matches!(error, LayoutError::InvalidParent(message) if message.contains("parent cycle"))
+    );
 }
 
 #[test]
@@ -104,4 +158,8 @@ fn curated_dagre_api_lays_out_flat_graph() {
     assert!(a.y().is_some());
     assert!(b.x().is_some());
     assert!(b.y().is_some());
+    assert!(b.y().expect("node B has y") > a.y().expect("node A has y"));
+
+    let edge = graph.edge("A", "B").expect("edge A -> B exists");
+    assert!(edge.points().len() >= 2);
 }
